pyproject.toml
[project]
name = "auto-trade"
version = "0.1.0"
description = "A股量化交易系统 - 回测与信号生成"
requires-python = ">=3.10"
dependencies = [
    "polars>=1.0.0",
    "akshare>=1.14.0",
    "pyarrow>=15.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
]
